refactor: PlaceInfo를 application/dto로 이동
outbound 어댑터 패키지에 있던 PlaceInfo 레코드를 application/dto로 이동.
HotelSearchController가 outbound DTO를 직접 노출하던 문제 해결.
/supply/hotels/place-info 엔드포인트에 @Operation 문서 추가.
